[BEAM-1871] Move functionality from AvroUtils to AvroSource hiding visibility of internal method, also remove deprecated method.
